+++ src/sound/streams.h
@@ -15,18 +15,18 @@ void streams_sh_stop(void);
 void streams_sh_update(void);
 
 int stream_init(const char *name,int default_mixing_level,
-		int sample_rate,
+		double sample_rate,
 		int param,void (*callback)(int param,INT16 *buffer,int length));
 int stream_init_float(const char *name,int default_mixing_level,
-		int sample_rate,
+		double sample_rate,
 		int param,void (*callback)(int param,INT16 *buffer,int length),int is_float);
 int stream_init_multi(int channels,const char **names,const int *default_mixing_levels,
-		int sample_rate,
+		double sample_rate,
 		int param,void (*callback)(int param,INT16 **buffer,int length));
 void stream_update(int channel,int min_interval);	/* min_interval is in usec */
 
-void stream_set_sample_rate(int channel, int sample_rate);
-int stream_get_sample_rate(int channel);
+void stream_set_sample_rate(int channel, double sample_rate);
+double stream_get_sample_rate(int channel);
 
 #ifdef __cplusplus
 }
